✨(frontend) add csrftoken to the frontend context
Action requests such as POST and PUT requests get blocked by the
Django CSRF protection when we are using Session middleware to
authenticate users on API calls. We need to make one such request
for the user enrollment feature we're building.

We can avoid this issue by passing the csrf token through a
dedicated header in our API requests. To do this, we must first
add the csrftoken to the context we're passing to our frontend.